+++ docs/.vuepress/theme/components/DialtoneLogo.vue
@@ -1,0 +1,21 @@
+<template>
+  <router-link
+    class="d-btn d-py0"
+    title="Go back to the homepage"
+    to="/"
+  >
+    <img
+      class="d-h42"
+      :alt="brandLogoAlt"
+      :src="brandLogo"
+    >
+  </router-link>
+</template>
+
+<script setup>
+import { useThemeLocaleData } from '@vuepress/plugin-theme-data/client';
+import { useSiteLocaleData } from '@vuepress/client';
+
+const brandLogo = useThemeLocaleData().value.logo;
+const brandLogoAlt = useSiteLocaleData().value.title;
+</script>